Modify JSON output of TopicModelBolt

+++ locationTopicModelTopology/src/main/java/uniko/west/topology/datatypes/MessageLocationPrediction.java
@@ -11,40 +11,39 @@ import java.util.Map.Entry;
 
 public class MessageLocationPrediction {
 	public HashMap<String, Object> message;
-	public HashMap<String, Double> locationProbabilityMap;
+	public HashMap<String, Float> locationProbabilityMap;
 
 	public MessageLocationPrediction(HashMap<String, Object> message) {
 		this.message = message;
-		locationProbabilityMap = new HashMap<String, Double>();
+		locationProbabilityMap = new HashMap<String, Float>();
 	}
 
-	public void addLocationProbability(String location, Double probability) {
+	public void addLocationProbability(String location, Float probability) {
 		locationProbabilityMap.put(location, probability);
 	}
 
 	public String getTopLocationsWithProbability(int numberOfLocations) {
 		// approach taken form: http://stackoverflow.com/a/13913206/2174538
-		List<Entry<String, Double>> list = new LinkedList<Entry<String, Double>>(
+		List<Entry<String, Float>> list = new LinkedList<Entry<String, Float>>(
 				locationProbabilityMap.entrySet());
 
 		// Sorting the list based on values
-		Collections.sort(list, new Comparator<Entry<String, Double>>() {
-			public int compare(Entry<String, Double> o1,
-					Entry<String, Double> o2) {
+		Collections.sort(list, new Comparator<Entry<String, Float>>() {
+			public int compare(Entry<String, Float> o1, Entry<String, Float> o2) {
 				return o2.getValue().compareTo(o1.getValue());
 
 			}
 		});
 
 		// Maintaining insertion order with the help of LinkedList
-		Map<String, Double> sortedLocationProbabilityMap = new LinkedHashMap<String, Double>();
-		for (Entry<String, Double> entry : list) {
+		Map<String, Float> sortedLocationProbabilityMap = new LinkedHashMap<String, Float>();
+		for (Entry<String, Float> entry : list) {
 			sortedLocationProbabilityMap.put(entry.getKey(), entry.getValue());
 		}
 
 		String resultString = "[";
 		int addedLocations = 0;
-		for (Entry<String, Double> entry : sortedLocationProbabilityMap
+		for (Entry<String, Float> entry : sortedLocationProbabilityMap
 				.entrySet()) {
 			if (addedLocations == numberOfLocations) {
 				break;
